Refactor rda/types.hpp to rda/rda_types.hpp
