+++ wxdata/wxdata.cmake
@@ -110,7 +110,7 @@ set(HDR_WSR88D_RDA include/scwx/wsr88d/rda/clutter_filter_bypass_map.hpp
                    include/scwx/wsr88d/rda/performance_maintenance_data.hpp
                    include/scwx/wsr88d/rda/rda_adaptation_data.hpp
                    include/scwx/wsr88d/rda/rda_status_data.hpp
-                   include/scwx/wsr88d/rda/types.hpp
+                   include/scwx/wsr88d/rda/rda_types.hpp
                    include/scwx/wsr88d/rda/volume_coverage_pattern_data.hpp)
 set(SRC_WSR88D_RDA source/scwx/wsr88d/rda/clutter_filter_bypass_map.cpp
                    source/scwx/wsr88d/rda/clutter_filter_map.cpp
